backend/requirements.txt
Django==4.2.7
django-ninja==0.22.2
pydantic==2.5.0
python-multipart==0.0.6
uvicorn==0.24.0

django-import-export
xlwt 
openpyxl 
xlrd
django-cors-headers==4.8.0 
PyJWT==2.10.1